#31: feature_add_place 모듈 내 네트워크 관련 구현 수정
