+++ feature_add_place/src/main/java/com/jyp/feature_add_place/domain/SearchPlaceRepositoryInterface.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.flow.Flow
 abstract class SearchPlaceRepositoryInterface {
     abstract fun getSearchPlaceResult(
         placeName: String
-    ): Flow<List<SearchPlaceResultModel>>
+    ): Flow<Result<List<SearchPlaceResultModel>>>
 }
